Cleanup after build

+++ fscps.tools/internal/functions/invoke-fsccompile.ps1
@@ -673,36 +673,37 @@ function Invoke-FSCCompile {
         finally
         {
             try {
-                <#
-                if($SolutionBuildFolderPath)
+                if($settings.cleanupAfterBuild)
                 {
-                    if (Test-Path -Path $SolutionBuildFolderPath -ErrorAction SilentlyContinue)
+                    if($SolutionBuildFolderPath)
                     {
-                        Remove-Item -Path $SolutionBuildFolderPath -Recurse -Force -ErrorAction SilentlyContinue
+                        if (Test-Path -Path $SolutionBuildFolderPath -ErrorAction SilentlyContinue)
+                        {
+                            Remove-Item -Path $SolutionBuildFolderPath -Recurse -Force -ErrorAction SilentlyContinue
+                        }
+                    }
+                    if($NuGetPackagesPath)
+                    {
+                        if (Test-Path -Path $NuGetPackagesPath -ErrorAction SilentlyContinue)
+                        {
+                            Remove-Item -Path $NuGetPackagesPath -Recurse -Force -ErrorAction SilentlyContinue
+                        }
+                    }
+                    if($outputDir)
+                    {
+                        if (Test-Path -Path $outputDir -ErrorAction SilentlyContinue)
+                        {
+                            Remove-Item -Path $outputDir -Recurse -Force -ErrorAction SilentlyContinue
+                        }
+                    }
+                    if($tempCombinedPackage)
+                    {
+                        if (Test-Path -Path $tempCombinedPackage -ErrorAction SilentlyContinue)
+                        {
+                            Remove-Item -Path $tempCombinedPackage -Recurse -Force -ErrorAction SilentlyContinue
+                        }
                     }
                 }
-                if($NuGetPackagesPath)
-                {
-                    if (Test-Path -Path $NuGetPackagesPath -ErrorAction SilentlyContinue)
-                    {
-                        Remove-Item -Path $NuGetPackagesPath -Recurse -Force -ErrorAction SilentlyContinue
-                    }
-                }
-                if($outputDir)
-                {
-                    if (Test-Path -Path $outputDir -ErrorAction SilentlyContinue)
-                    {
-                        Remove-Item -Path $outputDir -Recurse -Force -ErrorAction SilentlyContinue
-                    }
-                }
-                if($tempCombinedPackage)
-                {
-                    if (Test-Path -Path $tempCombinedPackage -ErrorAction SilentlyContinue)
-                    {
-                        Remove-Item -Path $tempCombinedPackage -Recurse -Force -ErrorAction SilentlyContinue
-                    }
-                }
-                    #>
             }
             catch {
                 Write-PSFMessage -Level Verbose -Message "Cleanup warning: $($PSItem.Exception)"
